Remove 2014b support.
There is a potential so-called ``static TLS`` problem, which related to
loading libs. Since Matlab is trying to pre-loading some libs on start,
it might be a problem if you are going to load another lib later.
However, this is for sure a bug from earlier version of Matlab.  The
problem could be somewhere about OPENMP. Removing ``OpenMP`` is
considered. Or using another libomp might help.

Signed-off-by: Yimin <zhongym8903@gmail.com>