create PDEcs class for PDE constrained solver.
This example is a toy model for reconstruction of coefficient of PDE, though it is only 1 dimensional, however, evaluating gradient will be more or less O(n) complexity for a normal case. And it will be very slow in many cases.

For corase mesh, sometimes it cannot find the optimal since the problem maybe not convex or search direction is not good enough.