Add regularization to the test case.
It is very easy to prove the uniqueness for this simple case, however, the problem is
not very good, since it is EIT problem related, statbility is  very bad, and regularization is almost always needed.

without regularization coefficient beta, the algorithm will stop very early and cannot proceed. A very tiny beta will push it forward.