remove dg, the reason is following:
DG actually needs another mesh output and it is not nodal based, it is
basis-based.

each element with a specific order will restrict our dofs, not like CG,
we are dealling with vertex-valued matrix.

With high order N and M triangles, usually if mesh is fine enough, we
roughly can use Euler  F + V - E = 2 and 3F = 2E to get vertices number
~ M/2.

for 2D.

CG needs  M/2 + M * ((N + 2)*(N + 1)/2 - 3) = 0.5 * M * (N^2 + 3N -3)
DG needs M * ((N + 2) * (N + 1)/2) 

DG will need about twice the computation cost, accuracy largely depends
on the basis.

and there are penalty terms to handle, it is not realistic to integrate
this into the original femex code, mainly because it lives on a new mesh
ordering system, and it does not use vertex value either, which rules
out the practice use of our other modules.


Signed-off-by: Yimin <zhongym8903@gmail.com>